move App to components app folder
